feat: implement Chapter 15 - A Virtual Machine
Add complete stack-based bytecode virtual machine with arithmetic operations,
runtime execution, configurable tracing, and proper error handling for the
Lox interpreter implementation.

+++ src/value.zig
@@ -1,5 +1,5 @@
 const std = @import("std");
-pub const Value = u64;
+pub const Value = f64;
 
 pub fn print(value: Value) void {
     std.debug.print("{d}", .{value});
